+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support for processing location messages to `TyntecWhatsAppAdapter#processActivity`.
+- Support for sending location messages to `TyntecWhatsAppAdapter#sendActivities`.